[feature] Improve single sign-on for customer accounts
The authentication of TYPO3 frontend user and Magento customer provides
both directions. It doesn't matter what system is used for it, as well
as the password encryption format which TYPO3 or Magento uses.

related #9

Therefore the authentication is more separated from the user account
replication, for which a new TYPO3 content element is introduced, to hold
the replication links between sources and targets. The replication covers
additionally the primary billing address of Magento customers.