[task] Rename configuration class of the TYPO3 extension
Changes the TYPO3 extension class 'tx_typogento_configurationHelper' to
'tx_typogento_configuration'.

It should in the future to be used as a service rather than a utility,
and thus - to avoid the direct processing of arrays in different
places - as a central access layer for the configuration of the TYPO3
extension.