[feature] JavaScript rewriter and default setup in TYPO3 extension
Integrates a JavaScript for rewriting of HTML forms and Ajax requests
on client side, and provides a clear separation of HTTP POST parameters
to other TYPO3 extensions.

It also integrates a default setup as a static TypoScript template, to
provide a plug and play installation as a starting point for the user.